refactor(scripts): Refactor scripts organization
This new order will allow to add new clients for other APIs like
Youtube.
