+++ portfolio/src/main/webapp/index.html
@@ -1,80 +1,100 @@
 <!DOCTYPE html>
 <html class="has-navbar-fixed-bottom">
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <title>Ernesto Garcia | STEP Intern</title>
+    <link
+      rel="stylesheet"
+      href="https://cdn.jsdelivr.net/npm/bulma@0.9.0/css/bulma.min.css"
+    />
+    <link rel="stylesheet" href="./css/custom.css" />
+    <link rel="stylesheet" href="./css/font-awesome.min.css" />
+  </head>
 
-<head>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <title>Ernesto Garcia | STEP Intern</title>
-  <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bulma@0.9.0/css/bulma.min.css">
-  <link rel="stylesheet" href="./css/custom.css">
-  <link rel="stylesheet" href="./css/font-awesome.min.css">
-</head>
-
-<body>
-  <section class="section animate-1 animate--fadeInLeft">
-    <div class="container">
-      <figure class="image mb-4 is-128x128">
-        <img class="is-rounded" src="images/avatar.jpg">
-      </figure>
-      <h1 class="title">
-        Ernesto Garcia
-      </h1>
-      <p class="subtitle mb-1">
-        Google <strong>STEP</strong> Intern
-      </p>
-      <div class="has-text-info mt-2">
-        <a target="_blank" rel="noopener noreferrer" href="https://facebook.com/ernestogrcanw">
-          <span class="icon">
-            <i class="fa fa-facebook"></i>
-          </span>
-        </a>
-        <a target="_blank" rel="noopener noreferrer" href="https://instagram.com/ernestognw">
-          <span class="icon">
-            <i class="fa fa-instagram"></i>
-          </span>
-        </a>
-        <a target="_blank" rel="noopener noreferrer" href="https://github.com/ernestognw">
-          <span class="icon">
-            <i class="fa fa-github"></i>
-          </span>
-        </a>
-        <a target="_blank" rel="noopener noreferrer" href="https://linkedin.com/in/ernestognw">
-          <span class="icon">
-            <i class="fa fa-linkedin"></i>
-          </span>
-        </a>
+  <body>
+    <section class="section animate-1 animate--fadeInLeft">
+      <div class="container">
+        <figure class="image mb-4 is-128x128">
+          <img class="is-rounded" src="images/avatar.jpg" />
+        </figure>
+        <h1 class="title">
+          Ernesto Garcia
+        </h1>
+        <p class="subtitle mb-1">Google <strong>STEP</strong> Intern</p>
+        <div class="has-text-info mt-2">
+          <a
+            target="_blank"
+            rel="noopener noreferrer"
+            href="https://facebook.com/ernestogrcanw"
+          >
+            <span class="icon">
+              <i class="fa fa-facebook"></i>
+            </span>
+          </a>
+          <a
+            target="_blank"
+            rel="noopener noreferrer"
+            href="https://instagram.com/ernestognw"
+          >
+            <span class="icon">
+              <i class="fa fa-instagram"></i>
+            </span>
+          </a>
+          <a
+            target="_blank"
+            rel="noopener noreferrer"
+            href="https://github.com/ernestognw"
+          >
+            <span class="icon">
+              <i class="fa fa-github"></i>
+            </span>
+          </a>
+          <a
+            target="_blank"
+            rel="noopener noreferrer"
+            href="https://linkedin.com/in/ernestognw"
+          >
+            <span class="icon">
+              <i class="fa fa-linkedin"></i>
+            </span>
+          </a>
+        </div>
+        <div class="main--description mt-6">
+          <p class="is-size-4">
+            Hello!
+          </p>
+          <p class="mt-2">
+            My name is Ernesto Garcia, and
+            <strong>I'm an undegrad CS student</strong> from Monterrey Institute
+            Of Technology and Higher Education based in Monterrey.
+          </p>
+          <p class="mt-2">
+            <strong>I'm a huge Node.js/Javascript fan</strong>, and I love to
+            create complex and challenging web and mobile applications using
+            MERN stack technologies. I've been doing fullstack web development
+            for nearly <strong>3 years</strong>, and I'm excited to start
+            @Google.
+          </p>
+        </div>
       </div>
-      <div class="main--description mt-6">
-        <p class="is-size-4">
-          Hello!
+    </section>
+    <section class="section animate-2 animate--fadeInLeft">
+      <div class="container">
+        <p class="title mb-4 is-size-4">
+          My projects
         </p>
-        <p class="mt-2">
-          My name is Ernesto Garcia, and <strong>I'm an undegrad CS student</strong> from Monterrey Institute Of
-          Technology and Higher
-          Education based in Monterrey.
-        </p>
-        <p class="mt-2">
-          <strong>I'm a huge Node.js/Javascript fan</strong>, and I love to create complex and challenging web and
-          mobile applications using MERN stack technologies.
-          I've been doing fullstack web development for nearly <strong>3 years</strong>, and I'm excited to start
-          @Google.
-        </p>
+        <div id="repositories" class="repositories--list"></div>
+        <button
+          id="more-repos-button"
+          class="button is-link is-light mt-5 block-x-center"
+        >
+          Load more...
+        </button>
       </div>
-    </div>
-  </section>
-  <section class="section animate-2 animate--fadeInLeft">
-    <div class="container">
-      <p class="title mb-4 is-size-4">
-        My projects
-      </p>
-      <div id="repositories" class="repositories--list">
-
-      </div>
-      <button id="more-repos-button" class="button is-link is-light mt-5 block-x-center">Load more...</button>
-    </div>
-  </section>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.27.0/moment.min.js"></script>
-  <script src="./scripts/github.js"></script>
-</body>
-
+    </section>
+    <script src="https://apis.google.com/js/api.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.27.0/moment.min.js"></script>
+    <script type="module" src="./scripts/index.js"></script>
+  </body>
 </html>
